requirements.txt
playwright
pandas
xlsxwriter
openpyxl
rich
streamlit
python-dotenv
